correct type

+++ src/root.zig
@@ -5,7 +5,7 @@ pub const Self = @This();
 
 allocator: std.mem.Allocator,
 url: []const u8,
-handshake_timeout: u64,
+handshake_timeout: u32,
 max_size: u32,
 buffer_size: u32,
 max_retries: u32,
@@ -13,7 +13,7 @@ client: ?wss.Client = null,
 
 pub const Opts = struct {
     url: []const u8,
-    handshake_timeout: u64 = 10000,
+    handshake_timeout: u32 = 10000,
     max_size: u32 = 4096,
     buffer_size: u32 = 1024,
     max_retries: u32 = 10,
